==> Modulo #4- Creacion de Portafolio y Conclusion/Markdowns/01_Creacion del Portafolio.Rmd <==
---
title: "Creacion del Portafolio"
subtitle: 'R con Enfoque Financiero: AulaABA '
date: "2023-06-06"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## De finicion del mix de activos en nuestro portafolio

Luego de importar nuestro archivo de retornos mensuales definimos los pesos de nuestro mix de acciones dentro de nuestro portafolio y le especificamos que queremos el rebalanceo de nuestro portafolio sea cada ano.

El rebalanceo de un portafolio anual es el proceso de ajustar las ponderaciones de los activos dentro de un portafolio para mantener la asignación de activos deseada en un horizonte de tiempo específico, generalmente un año.

Cuando se construye un portafolio, se establece una asignación de pesos o porcentajes para cada activo o clase de activos con el objetivo de alcanzar ciertos objetivos de rendimiento y nivel de riesgo. Sin embargo, a medida que pasa el tiempo, el valor de los activos individuales puede cambiar y, como resultado, las ponderaciones iniciales se desvían de la asignación deseada.

```{r}
suppressWarnings(suppressPackageStartupMessages(library(PerformanceAnalytics)))
suppressWarnings(suppressPackageStartupMessages(library(quantmod)))
suppressWarnings(suppressPackageStartupMessages(library(tidyverse)))

retornos_mensuales_FAANG <- readRDS("retornos_mensuales_FAANG.rds")

n <- ncol(retornos_mensuales_FAANG)
pesos_iguales <- rep(1/n, n) ##Cada acción tendrá un peso inicial correspondiente al 20%

FAANG_EqualWeights <- Return.portfolio(
  R= retornos_mensuales_FAANG, 
  weights= pesos_iguales, 
  rebalance_on ="years", 
  verbose= TRUE ##Activamos el verbose para obtener una 
  #lista con diferentes parámetros relacionados al cálculo del retorno de este indice
)

##Obtenemos el objeto de retornos del portafolio
Retornos_FAAN_EqualWeights <- FAANG_EqualWeights$returns
colnames(Retornos_FAAN_EqualWeights) <- "FAAMG-EqualWeights"
head(FAANG_EqualWeights$contribution)
```

## Visualizamos el Peso de Cada accion al final del periodo

Es importante tener en cuenta que la periodicidad del rebalanceo y los pesos iniciales de cada acción se pueden modificar de acuerdo la estrategia de inversión seleccionada o un análisis técnico anterior como una optimización recurrente.

```{r}
##Graficamos la evolución de los pesos de cada una de las 
#acciones FAANG para tener una idea de como Return.portfolio hace
#los rebalanceos en los periodos especificados
PesosFinalDelPeriodo <- data.frame(FAANG_EqualWeights$EOP.Weight) 
PesosFinalDelPeriodo <- tibble::rownames_to_column(PesosFinalDelPeriodo, "fecha")
PesosFinalDelPeriodo2 <- gather(PesosFinalDelPeriodo, Acciones, Pesos, -fecha)
PesosFinalDelPeriodo2 %>%
  ggplot( aes(x=as.Date(fecha), y=Pesos, group=Acciones, color= Acciones)) +
    geom_line() + ggtitle("Evolución de los Pesos de las Acciones FAANG") + labs(x= "Fecha") +
          geom_vline(xintercept = c(
            as.Date("2016-01-01"),
            as.Date("2017-01-01"),
                                    as.Date("2018-01-01"),
                                    as.Date("2019-01-01"),
                                    as.Date("2020-01-01"),
                                    as.Date("2021-01-01"),
                                    as.Date("2022-01-01")
                                    ) , linetype="dotted",  color = "blue", linewidth=0.2) +
  theme_minimal()
```

```{r}
write_rds(FAANG_EqualWeights, "portafolio_FAANG_EW.rds")
```
