==> Modulo #4- Creacion de Portafolio y Conclusion/Markdowns/03_Evaluacion del Riesgo de un Portafolio.Rmd <==
---
title: "Evaluacion del Riesgo de un Portafolio"
subtitle: 'R con Enfoque Financiero: AulaABA '
output:
  pdf_document: default
  html_document: default
date: "2023-06-06"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Desviacion Estandar de los Portafolios

La desviacion estandar de los retornos mide como los retornos se encuentran distribuidos alrededor de su promedio. Si los retornos se encuentran dispersos, el portafolio tiene una mayor desviacion estandar y es visto como mas riesgoso y volatil.

Para hacer este calculo tomamos los siguientes pasos:

-   Calculamos la matriz de covarianza entre los rendimientos de las acciones. La covarianza mide cómo los rendimientos de las acciones se mueven juntas. Aqui podemos utilizar funciones de R, como cov() o cov2cor(), para calcular la matriz de covarianza.

-   Calculamos la varianza del portafolio. La varianza del portafolio se calcula multiplicando la matriz de covarianza por los pesos de las acciones y luego multiplicando el resultado por los pesos transpuestos. Es decir, varianza_del_portafolio = pesos_transpuestos \* matriz_de_covarianza \* pesos.

-   Calculamos la desviación estándar del portafolio. La desviación estándar es simplemente la raíz cuadrada de la varianza del portafolio. Podemos calcularla utilizando la función sqrt() en R.

```{r}

suppressWarnings(suppressPackageStartupMessages(library(PerformanceAnalytics)))
suppressWarnings(suppressPackageStartupMessages(library(quantmod)))
suppressWarnings(suppressPackageStartupMessages(library(tidyverse)))

FAANG <- readRDS("portafolio_FAANG_EW.rds")
retornos_mensuales_FAANG <- readRDS("retornos_mensuales_FAANG.rds")

n <- ncol(retornos_mensuales_FAANG)
pesos_iguales <- rep(1/n, n) ##Cada acción tendrá un peso inicial correspondiente al 20%

###Calculando manualmente la desviacion de un portafolio. 
matriz_covarianzas<- cov(retornos_mensuales_FAANG)
matriz_desviaciones_estandar <- sqrt(t(pesos_iguales) %*%
                                       matriz_covarianzas %*%
                                       pesos_iguales)
print(matriz_desviaciones_estandar)
```

## Desviacion Estandar en PerformanceAnalytics

```{r, echo=FALSE}
n <- ncol(retornos_mensuales_FAANG)
pesos_iguales <- rep(1/n, n) ##Cada acción tendrá un peso inicial correspondiente al 20%

FAANG_StdDev <-  StdDev(retornos_mensuales_FAANG,
                        weights = pesos_iguales)
FAANG_StdDev <- data.frame(Stock= "FAANG Port",
                           desviacion= FAANG_StdDev)
head(FAANG_StdDev)
```

## Visualizando las Desviaciones Estandar

```{r}
retornos_portf_faang <- FAANG$returns %>%
  data.frame(fecha= index(.)) %>%
  remove_rownames() 

desviacion_grafico <- sd(FAANG$returns)
promedio_grafico <- mean(FAANG$returns)

retornos_portf_faang <- retornos_portf_faang  |>
mutate(hist_color_rojo= 
ifelse(portfolio.returns < (promedio_grafico- desviacion_grafico),
       portfolio.returns, as.numeric(NA)), 
hist_color_verde= 
ifelse(portfolio.returns > (promedio_grafico+ desviacion_grafico),
       portfolio.returns, as.numeric(NA)),
hist_color_azul= 
ifelse(portfolio.returns > (promedio_grafico- desviacion_grafico) &
portfolio.returns < (promedio_grafico+ desviacion_grafico),
portfolio.returns, as.numeric(NA)))

retorno_FAANG <-  ggplot(data = retornos_portf_faang, aes(x = fecha)) +
geom_point(aes(y = hist_color_rojo*100), color= "red" )+
geom_point(aes(y = hist_color_verde*100), color= "green" )+
geom_point(aes(y = hist_color_azul*100), color= "blue" )+
geom_hline(yintercept =((promedio_grafico + desviacion_grafico)*100),
    color="purple", linetype="dotted") + 
geom_hline(yintercept =((promedio_grafico - desviacion_grafico)*100),
    color="purple", linetype="dotted") + 
scale_x_date(date_labels = "%b-%Y")  +
labs(title= "Evolucion del Retorno del Portafolio FAANG",
subtitle =  paste("Valores Mensuales de", min(retornos_portf_faang$fecha),
                 "a", max(retornos_portf_faang$fecha)), 
x= "Fecha",
y= "Retorno", 
caption= "Yahoo Finance") + 
theme_classic() + 
theme(legend.position = "none",
legend.title = element_blank()) + 
theme_classic()

suppressWarnings(plot(retorno_FAANG))
```

```{r}

precios_FAANG_vect <- c("AMZN", "META", "NFLX", "GOOGL", "AAPL")

retornos_mensuales_FAANG <- readRDS("retornos_mensuales_FAANG.rds")

retornos_mensuales_FAANG_largo <- retornos_mensuales_FAANG %>%
  data.frame(fecha= index(.)) %>%
  remove_rownames() %>% 
   pivot_longer(cols =all_of(precios_FAANG_vect),
    names_to = "Stock", values_to = "Retornos") 

retornos_mensuales_FAANG_largo <- retornos_mensuales_FAANG_largo %>%
  group_by(Stock) %>% 
  summarise(desviacion= sd(Retornos)) %>% 
  add_row(FAANG_StdDev)

```

```{r}
ggplot(data= retornos_mensuales_FAANG_largo, 
       aes(x= Stock, y= desviacion, color= Stock)) +
geom_point()   +
labs(title= "Desviacion del Portafolio FAANG vs. sus componentes",
   subtitle =  paste("Fecha de", min(retornos_portf_faang$fecha),
                     "a", max(retornos_portf_faang$fecha)), 
   x= "Fecha",
   y= "Desviacion", 
   caption= "Yahoo Finance") + 
theme(legend.position = "none",
    legend.title = element_blank()) + 
theme_classic()

```
